+++ packages/media/media-store/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/media-store
 
+## 6.2.1
+- [patch] Updated dependencies [927ae63](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/927ae63)
+  - @atlaskit/media-test-helpers@18.0.0
+
 ## 6.2.0
 - [minor] expose new context.collection methods [6e1d642](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6e1d642)